shell-ai/pyproject.toml
[project]
name = "shell-ai"
version = "0.1.0"
description = "A stylish multi-model AI assistant for your shell"
requires-python = ">=3.11"
dependencies = [
    "rich>=13.7",
    "prompt-toolkit>=3.0",
    "httpx>=0.27",
    "openai>=1.30",
    "anthropic>=0.25",
    "google-genai>=1.0",
    "toml>=0.10",
]

[project.scripts]
shai = "shell_ai.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
